changed visual appearance and moved brainstorm related files to a new folder
