+++ WeVoteWeTalk/we_vote_we_talk/lib/Brainstorm/Idea.dart
